requirements.txt
absl-py==1.0.0
asgiref==3.4.1
beautifulsoup4==4.10.0
boto3==1.21.18
botocore==1.24.18
certifi==2020.6.20
charset-normalizer==2.0.12
click==8.0.3
dataclasses==0.8
Django==3.2.12
feedparser==6.0.8
filelock==3.4.1
gensim==4.1.2
huggingface-hub==0.4.0
idna==3.3
importlib-metadata==4.8.3
jmespath==0.10.0
joblib==1.1.0
nltk==3.6.7
numpy==1.19.5
openai==0.8.0
packaging==21.3
Pillow==8.4.0
pyemd==0.5.1
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
regex==2022.1.18
requests==2.27.1
rouge-score==0.0.4
s3transfer==0.5.2
sacremoses==0.0.47
scikit-learn==0.24.2
scipy==1.5.4
sentence-transformers==2.2.0
sentencepiece==0.1.96
sgmllib3k==1.0.0
six==1.16.0
smart-open==5.2.1
soupsieve==2.3.1
sqlparse==0.4.2
threadpoolctl==3.1.0
tokenizers==0.11.4
torch==1.10.2
torchvision==0.11.3
tqdm==4.62.3
transformers==4.16.2
typing_extensions==4.1.1
urllib3==1.26.8
zipp==3.6.0
